+++ packages/html/src/chat/templates/chat-empty.tsx
@@ -1,0 +1,6 @@
+import { Chat } from "..";
+
+export const ChatEmpty = (props) => (
+    <Chat empty
+     {...props} />
+);